Move call to check_access to call_if
