+++ src/api-service/__app__/onefuzzlib/endpoint_authorization.py
@@ -4,20 +4,71 @@
 # Licensed under the MIT License.
 
 import logging
+import urllib
 from typing import Callable, Optional
 from uuid import UUID
 
 import azure.functions as func
+from azure.functions import HttpRequest
 from memoization import cached
 from onefuzztypes.enums import ErrorCode
 from onefuzztypes.models import Error, UserInfo
 
 from .azure.creds import get_scaleset_principal_id
+from .azure.group_membership import create_group_membership_checker
 from .config import InstanceConfig
 from .request import not_ok
+from .request_access import RequestAccess
 from .user_credentials import parse_jwt_token
 from .workers.pools import Pool
 from .workers.scalesets import Scaleset
+
+
+@cached(ttl=60)
+def get_rules() -> Optional[RequestAccess]:
+    config = InstanceConfig.fetch()
+    if config.api_access_rules:
+        return RequestAccess.build(config.api_access_rules)
+    else:
+        return None
+
+
+def check_access(req: HttpRequest) -> Optional[Error]:
+    rules = get_rules()
+
+    # Noting to enforce if there are no rules.
+    if not rules:
+        return None
+
+    path = urllib.parse.urlparse(req.url).path
+    rule = rules.get_matching_rules(req.method, path)
+
+    # No restriction defined on this endpoint.
+    if not rule:
+        return None
+
+    member_id = UUID(req.headers["x-ms-client-principal-id"])
+
+    try:
+        membership_checker = create_group_membership_checker()
+        allowed = membership_checker.is_member(rule.allowed_groups_ids, member_id)
+        if not allowed:
+            logging.error(
+                "unauthorized access: %s is not authorized to access in %s",
+                member_id,
+                req.url,
+            )
+            return Error(
+                code=ErrorCode.UNAUTHORIZED,
+                errors=["not approved to use this endpoint"],
+            )
+    except Exception as e:
+        return Error(
+            code=ErrorCode.UNAUTHORIZED,
+            errors=["unable to interact with graph", str(e)],
+        )
+
+    return None
 
 
 @cached(ttl=60)
@@ -120,6 +171,11 @@ def call_if(
     allow_user: bool = False,
     allow_agent: bool = False
 ) -> func.HttpResponse:
+
+    access = check_access(req)
+    if isinstance(access, Error):
+        return access
+
     token = parse_jwt_token(req)
     if isinstance(token, Error):
         return not_ok(token, status_code=401, context="token verification")
